ana menü çok büyümüşdü componentlere ayırıldı
